fix:core: Set confirmedRequests as Integer in Internship; fix:core: Set confirmed_requests as Integer in schema.sql; fix:core: Correct insert students in data.sql; fix:core: Add transactional in InternshipServiceImpl; fix:core: Setting confirmedRequests in InternshipServiceImpl::add;

+++ t-core/src/main/java/com/pavbatol/talentium/internship/dto/InternshipDtoFull.java
@@ -32,7 +32,7 @@ public class InternshipDtoFull {
 
     Integer participantLimit;
 
-    Long confirmedRequests;
+    Integer confirmedRequests;
 
     LocalDateTime createdOn;
